wx-gov-api/pyproject.toml
[project]
name = "watsonx-governance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "python-dotenv>=1.0.1",
    "fastapi>=0.116.1",
    "ibm-watsonx-gov[metrics]>=1.4.2",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
    "nest-asyncio>=1.6.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-prometheus>=0.48b0",
    "prometheus-client>=0.20.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
